@@ -3,6 +3,7 @@
     "version": "1.0.0",
     "description": "",
     "scripts": {
+        "prebuild": "rimraf dist",
         "start": "npm run serve",
         "build": "npm run build-ts && npm run eslint",
         "serve": "node dist/app.js",
@@ -11,8 +12,8 @@
         "test": "jest --forceExit ---verbose",
         "prettier-watch": "onchange 'src/*' -- prettier --write {{changed}}",
         "watch-test": "npm run test -- --watchAll",
-        "build-ts": "tsc",
-        "watch-ts": "tsc -w",
+        "build-ts": "npm run prebuild && tsc",
+        "watch-ts": "npm run prebuild && tsc -w",
         "eslint": "eslint src/*",
         "eslint-fix": "eslint src/* --fix",
         "debug": "npm run build && npm run watch-debug",
@@ -30,7 +31,7 @@
         "@types/jest": "^26.0.24",
         "@types/jsonwebtoken": "^8.5.4",
         "@types/morgan": "^1.9.3",
-        "@types/node": "^16.3.1",
+        "@types/node": "^16.4.2",
         "@types/passport": "^1.0.7",
         "@types/passport-facebook-token": "^0.4.35",
         "@types/passport-google-oauth": "^1.0.42",
@@ -38,10 +39,10 @@
         "@types/redis": "^2.8.31",
         "@types/speakeasy": "^2.0.6",
         "@types/ua-parser-js": "^0.7.36",
-        "@typescript-eslint/eslint-plugin": "^4.28.2",
-        "@typescript-eslint/parser": "^4.28.2",
+        "@typescript-eslint/eslint-plugin": "^4.28.4",
+        "@typescript-eslint/parser": "^4.28.4",
         "concurrently": "^6.2.0",
-        "eslint": "^7.30.0",
+        "eslint": "^7.31.0",
         "eslint-config-google": "^0.14.0",
         "eslint-config-prettier": "^8.3.0",
         "eslint-plugin-prettier": "^3.4.0",
@@ -50,11 +51,12 @@
         "nodemon": "^2.0.12",
         "onchange": "^7.1.0",
         "prettier": "^2.3.2",
-        "ts-jest": "^27.0.3",
+        "testcontainers": "^7.18.1",
+        "ts-jest": "^27.0.4",
         "typescript": "^4.3.5"
     },
     "dependencies": {
-        "@sentry/node": "^6.8.0",
+        "@sentry/node": "^6.10.0",
         "amqplib": "^0.8.0",
         "bcrypt": "^5.0.1",
         "blake3": "^2.1.4",
@@ -64,7 +66,7 @@
         "dotenv": "^10.0.0",
         "express": "^4.17.1",
         "express-async-handler": "^1.1.4",
-        "express-validator": "^6.12.0",
+        "express-validator": "^6.12.1",
         "helmet": "^4.6.0",
         "http-status-codes": "^2.1.4",
         "jsonwebtoken": "^8.5.1",
@@ -79,6 +81,7 @@
         "recaptcha2": "^1.3.3",
         "redis": "^3.1.2",
         "reflect-metadata": "^0.1.13",
+        "rimraf": "^3.0.2",
         "speakeasy": "^2.0.0",
         "tsyringe": "^4.6.0",
         "typeorm": "^0.2.34",
