@@ -1,39 +1,40 @@
 import { NextFunction, Request, Response } from "express";
-import { singleton } from "tsyringe";
+import { inject, singleton } from "tsyringe";
 
-import { StaleRefreshTokenException } from "../exceptions/exceptions";
 import { RoleManager } from "../managers/roleManager";
-import { SessionManager } from "../managers/sessionManager";
+import { BaseSessionManager } from "../managers/session/baseSessionManager";
 import { Session } from "../models/session";
 import { JwtService } from "../services/jwtService";
 import { forwardInternalError } from "../utils/expressUtils";
-import { REFRESH_TOKEN_COOKIE_NAME } from "../utils/globalConsts";
+import { SESSION_COOKIE_NAME } from "../utils/globalConsts";
 import * as errors from "./commonErrors";
 
 @singleton()
 export default class TokenController {
-    constructor(private _roleManager: RoleManager, private _sessionManager: SessionManager, private _jwtService: JwtService) {}
+    constructor(
+        private _roleManager: RoleManager,
+        @inject(BaseSessionManager.name) private _sessionManager: BaseSessionManager,
+        private _jwtService: JwtService
+    ) {}
 
     public async refreshAccessToken(req: Request, res: Response, next: NextFunction) {
-        const refreshToken = req.cookies[REFRESH_TOKEN_COOKIE_NAME];
+        const sessionCookie = req.cookies[SESSION_COOKIE_NAME];
 
         let session: Session;
         try {
-            session = await this._sessionManager.refreshSession(refreshToken, req.ip);
+            session = await this._sessionManager.refreshJwt(sessionCookie, req.ip);
         } catch (error) {
-            if (error instanceof StaleRefreshTokenException) {
-                return errors.staleRefreshToken(next);
-            }
             return forwardInternalError(next, error);
         }
 
         if (!session) {
+            res.clearCookie(SESSION_COOKIE_NAME);
             return errors.sessionDoesNotExist(next);
         }
 
         const roles = await this._roleManager.getRoles(session.userId);
-        const accessToken = this._jwtService.issueAccessToken(refreshToken, session.userId, roles);
+        const accessToken = this._jwtService.issueAccessToken(sessionCookie, session.userId, roles);
 
-        res.json({ accessToken: accessToken });
+        res.json({ accessToken });
     }
 }
