@@ -39,7 +39,7 @@ export default class ExternalUserController extends UserController {
             return;
         }
 
-        this.sendTokens(req, res, userId);
+        this.respondWithSessionOrJwt(req, res, userId);
     }
 
     public async finishRegistration(req: Request, res: Response, next: NextFunction) {
@@ -55,7 +55,7 @@ export default class ExternalUserController extends UserController {
 
         userId = await this.register(req.body, tokenData.id, tokenData.email, tokenData.provider);
 
-        this.sendTokens(req, res, userId);
+        this.respondWithSessionOrJwt(req, res, userId);
     }
 
     private async register(body: RequestBody, externalId: string, email: string, provider: ExternalLoginProvider): Promise<string> {
