@@ -2,7 +2,7 @@ import { BaseEntity, Column, CreateDateColumn, Entity, PrimaryColumn } from "typ
 
 @Entity({ name: "user" })
 export class UserEntity extends BaseEntity {
-    constructor(id: string, username?: string) {
+    constructor(id: string, username: string) {
         super();
         this.id = id;
         this.username = username;
@@ -13,6 +13,9 @@ export class UserEntity extends BaseEntity {
 
     @Column()
     username: string; // helpful with finding username duplicates between external and local accounts
+
+    @Column("text", { array: true, default: "{}" })
+    sessionIds: string[];
 
     @CreateDateColumn({ type: "timestamp" })
     createdAt: Date;
