@@ -1,12 +1,14 @@
 import { singleton } from "tsyringe";
-import { getRepository } from "typeorm";
+import { Connection } from "typeorm";
 
 import { ExternalLoginEntity, ExternalLoginProvider } from "../dal/entities/externalLoginEntity";
 import { guardNotUndefinedOrNull } from "../utils/guardClauses";
 
 @singleton()
 export class ExternalLoginManager {
-    private _repo = getRepository(ExternalLoginEntity);
+    private _repo = this._connection.getRepository(ExternalLoginEntity);
+
+    constructor(private _connection: Connection) {}
 
     public async create(userId: string, externalUserId: string, email: string, provider: ExternalLoginProvider) {
         const entity = new ExternalLoginEntity(userId, externalUserId, email, provider);
@@ -16,7 +18,7 @@ export class ExternalLoginManager {
     public async getUserId(externalUserId: string, provider: ExternalLoginProvider): Promise<string> {
         guardNotUndefinedOrNull(externalUserId);
 
-        const entity = await this._repo.findOne({ externalUserId: externalUserId, provider: provider }); // search also by provider to make sure there aren't two different users across platforms with same user id
+        const entity = await this._repo.findOne({ externalUserId, provider }); // search also by provider to make sure there aren't two different users across platforms with same user id
         if (!entity) {
             return null;
         }
