@@ -1,18 +1,13 @@
-import nameof from "../utils/nameof";
-import { Expirable } from "./expirable";
-
-export class Session extends Expirable {
-    private _token: string;
+export class Session {
+    private _id: string;
     private _userId: string;
     private _lastUseAt: Date;
     private _createIp: string;
     private _lastRefreshIp: string;
     private _createdAt: Date;
 
-    constructor(token: string, userId: string, lastUseAt: Date, createIp: string, lastRefreshIp: string, createdAt: Date) {
-        super(nameof<Session>("lastUseAt"));
-
-        this._token = token;
+    constructor(id: string, userId: string, lastUseAt: Date, createIp: string, lastRefreshIp: string, createdAt: Date) {
+        this._id = id;
         this._userId = userId;
         this._lastUseAt = lastUseAt;
         this._createIp = createIp;
@@ -20,8 +15,8 @@ export class Session extends Expirable {
         this._createdAt = createdAt;
     }
 
-    public get token(): string {
-        return this._token;
+    public get id(): string {
+        return this._id;
     }
 
     public get userId(): string {
