@@ -1,13 +1,17 @@
 import { singleton } from "tsyringe";
 
+import { guardNotUndefinedOrNull } from "../utils/guardClauses";
+
 // source: https://github.com/disposable/disposable
 @singleton()
 export class SpamProtector {
     private _list: Record<string, boolean>;
 
     public isDisallowedEmail(email: string): boolean {
+        guardNotUndefinedOrNull(email);
+
         const parts = email.split("@");
-        if (parts.length !== 2) {
+        if (parts.length !== 2 || !parts[0] || !parts[1]) {
             throw new Error("Invalid email address: " + email);
         }
         const domain = parts[1].toLowerCase();
