@@ -1,5 +1,5 @@
-import { REFRESH_TOKEN_COOKIE_NAME } from "../../utils/globalConsts";
+import { SESSION_COOKIE_NAME } from "../../utils/globalConsts";
 
 export interface RequestCookies {
-    [REFRESH_TOKEN_COOKIE_NAME]?: string;
+    [SESSION_COOKIE_NAME]?: string;
 }
