@@ -2,7 +2,7 @@ export const USER_ID_LENGTH = 14;
 export const PASSWORD_RESET_CODE_LENGTH = 32;
 export const MFA_LOGIN_TOKEN_LENGTH = 32;
 export const ONE_TIME_PASS_LENGTH = 6;
-export const JWT_ID_LENGTH = 5;
-export const REFRESH_TOKEN_LENGTH = 22;
+export const JWT_ID_LENGTH = 6;
+export const SESSION_ID_LENGTH = 22;
 export const CONFIRMATION_CODE_LENGTH = 6;
-export const REFRESH_TOKEN_COOKIE_NAME = "refresh-token";
+export const SESSION_COOKIE_NAME = "session-id";
