@@ -1,7 +1,8 @@
+import { NullOrUndefinedException } from "../exceptions/exceptions";
 import { isNullOrUndefined } from "./isNullOrUndefined";
 
 export function guardNotUndefinedOrNull(value: any) {
     if (isNullOrUndefined(value)) {
-        throw new Error("Value is null or undefined.");
+        throw new NullOrUndefinedException("Value is null or undefined.");
     }
 }
