@@ -1,17 +1,16 @@
-import "reflect-metadata";
-
+import { NullOrUndefinedException } from "../../src/exceptions/exceptions";
 import { guardNotUndefinedOrNull } from "../../src/utils/guardClauses";
 
 describe("guardNotUndefinedOrNull()", () => {
-    it("should throw error if argument is null", async () => {
-        expect(() => guardNotUndefinedOrNull(null)).toThrowError();
+    it("should throw error if argument is null", () => {
+        expect(() => guardNotUndefinedOrNull(null)).toThrowError(NullOrUndefinedException);
     });
 
-    it("should throw error if argument is undefined", async () => {
-        expect(() => guardNotUndefinedOrNull(undefined)).toThrowError();
+    it("should throw error if argument is undefined", () => {
+        expect(() => guardNotUndefinedOrNull(undefined)).toThrowError(NullOrUndefinedException);
     });
 
-    it("should not throw error if argument is not null nor undefined", async () => {
+    it("should not throw error if argument is not null nor undefined", () => {
         guardNotUndefinedOrNull("1");
     });
 });
